electric engine hydrazine and variable efficiency

+++ FNPlugin/FNModuleRadiation.cs
@@ -52,7 +52,7 @@ namespace FNPlugin {
 					break;
 				}
 				//rad = VanAllen.getBeltAntiparticles (cur_ref_body.flightGlobalsIndex, (float) (Vector3d.Distance(FlightGlobals.ship_position,cur_ref_body.transform.position)-cur_ref_body.Radius), 0.0f);
-                rad = (float) VanAllen.getRadiationLevel(cur_ref_body.flightGlobalsIndex, (Vector3d.Distance(FlightGlobals.ship_position, cur_ref_body.transform.position) - cur_ref_body.Radius), 0.0);
+                rad = VanAllen.getRadiationLevel(cur_ref_body.flightGlobalsIndex, (Vector3d.Distance(FlightGlobals.ship_position, cur_ref_body.transform.position) - cur_ref_body.Radius), 0.0);
 				if (cur_ref_body.flightGlobalsIndex == PluginHelper.REF_BODY_KERBOL) {
 					rad = rad * 1e6;
 				}
